+++ example/lib/home_page.dart
@@ -8,15 +8,43 @@ class HomePage extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       body: Column(
-        children: <Widget>[
-          const Text('Examples availables'),
-          const SizedBox(
+        children: const <Widget>[
+          Text('Examples availables'),
+          SizedBox(
             height: 16,
           ),
-          Text(
-            defaultUserModel.toString(),
+          _ListTile(
+            label: 'UserModel',
+            model: defaultUserModel,
+          ),
+          _ListTile(
+            label: 'AddressModel',
+            model: defaultAddressModel,
           ),
         ],
+      ),
+    );
+  }
+}
+
+class _ListTile extends StatelessWidget {
+  const _ListTile({
+    required this.label,
+    required this.model,
+  });
+
+  final String label;
+  final Model model;
+
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      title: Text(
+        label,
+        style: Theme.of(context).textTheme.titleMedium,
+      ),
+      subtitle: Text(
+        model.toString(),
       ),
     );
   }